Create anti-vowel.py
Question: 
Is there a more efficient/cleaner way to do this? Could I make a list and check if the entered string has any of the values from the list instead listing out each upper and lower vowel one by one?
The line breaks are where they because it looked nicer in the system I was using. 
The hint codecademy gave me is "To check to see if c is a vowel, you can do: c in "aeiouAEIOU"." but I don't know how to implement that. Maybe check if it char is in "aeio...." and return False but not True?